requirements.txt
Flask==3.0.0
tensorflow
opencv-python-headless
numpy
pillow
